+++ addOns/graphql/src/test/java/org/zaproxy/addon/graphql/InlineInjectorUnitTest.java
@@ -144,6 +144,14 @@ class InlineInjectorUnitTest extends TestUtils {
     }
 
     @Test
+    void injectInputObjectArgumentCorrectlyEscaped() {
+        String query = "{sqlInjection(expression: \"1\")}";
+        String sqliPaylaod = "\"or 1=1--";
+        String expectedQuery = "{sqlInjection(expression:\"\\\"or 1=1--\")}";
+        assertEquals(expectedQuery, injector.inject(query, "sqlInjection.expression", sqliPaylaod));
+    }
+
+    @Test
     void injectInlineFragmentArguments() {
         String query = "{...on field1{name(id:1)}}";
         String expectedQuery = "{...on field1{name(id:55)}}";