+++ addOns/graphql/CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Changed
 - Maintenance changes.
 
+### Fixed
+- Fix graphql parameter injection in URL query.
+
 ## [0.22.0] - 2023-12-19
 ### Added
 - Fingerprinting check for the GraphQL.NET engine.